project/requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
matplotlib
joblib
PyYAML
# Optional:
# xgboost
# lightgbm
